chore - refactored method to use single function export
